+++ build.gradle.kts
@@ -37,6 +37,8 @@ dependencies {
     implementation ("org.springframework.boot:spring-boot-starter-graphql")
     implementation("com.graphql-java-kickstart:graphql-spring-boot-starter:15.1.0")
     implementation("com.graphql-java-kickstart:graphiql-spring-boot-starter:11.1.0")
+    implementation ("io.springfox:springfox-boot-starter:3.0.0")
+
 
 //	util
     implementation("net.datafaker:datafaker:2.1.0")
