--- src/main/api/model/dto/TestDto.kt
@@ -1,10 +1,0 @@
-package com.daelim.springtest.main.api.model.dto
-
-data class TestDto(
-    val id: String,
-    val age: Int,
-)
-
-data class TestDtoRequest(
-    val age: Int,
-)
