+++ src/main/kotlin/api/config/SwaggerConfig.kt
@@ -1,0 +1,20 @@
+package api.config
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import springfox.documentation.builders.PathSelectors
+import springfox.documentation.builders.RequestHandlerSelectors
+import springfox.documentation.spi.DocumentationType
+import springfox.documentation.spring.web.plugins.Docket
+
+@Configuration
+class SwaggerConfig {
+    @Bean
+    fun api(): Docket {
+        return Docket(DocumentationType.SWAGGER_2)
+            .select()
+            .apis(RequestHandlerSelectors.basePackage("api")) // "api"는 컨트롤러가 위치한 패키지를 지정합니다.
+            .paths(PathSelectors.any())
+            .build()
+    }
+}
