+++ src/main/kotlin/com/daelim/springtest/main/api/model/dto/TestDto.kt
@@ -1,0 +1,11 @@
+package com.daelim.springtest.main.api.model.dto
+
+data class LottoNumbers(val numbers: List<Int>)
+
+data class GameResult(
+    val matchedNumbers: Int,
+    val isBonusMatched: Boolean,
+    val prize: String
+)
+
+data class WinningNumbers(val numbers: List<Int>)
